+++ app/router.js
@@ -6,6 +6,7 @@ const Router = Ember.Router.extend({
 });
 
 Router.map(function() {
+  this.route('sound-board', {path: '/'});
   this.route('about');
 });
