+++ app/templates/sound-board.hbs
@@ -1,0 +1,13 @@
+<div class="container">
+	{{#each pathAndLabelRow in audioPaths}}
+		<div class="row">
+			{{#each pathAndLabel in pathAndLabelRow}}
+				{{#sound-button audioFilePath=pathAndLabel.path}}
+					{{pathAndLabel.label}}
+				{{/sound-button	}}
+			{{/each}}
+		</div>
+	{{/each}}
+</div>
+
+<div class="play-random" {{action "playRandom"}}>Zoo Flesh</div>
